refactor: 优化项目目录结构
- 消息组件: Messages/ → chat/messages/
- 设置组件: Settings.vue → settings/Settings.vue
- Markdown 工具: components/Markdown.ts → utils/markdown.ts
- 统一目录命名: 使用小写 (inputs, settings)
- 更新所有导入路径
- 完善文档说明新的目录结构
