+++ src/App.vue
@@ -7,8 +7,8 @@ import ChatMessages from './components/chat/ChatMessages.vue'
 import SystemPrompt from './components/chat/SystemPrompt.vue'
 import CompactModelSelector from './components/common/CompactModelSelector.vue'
 import Sidebar from './components/common/Sidebar.vue'
-import TextInput from './components/Inputs/TextInput.vue'
-import Settings from './components/Settings.vue'
+import TextInput from './components/inputs/TextInput.vue'
+import Settings from './components/settings/Settings.vue'
 import { applyDarkModeToDocument, syncSystemDarkMode } from './composables/useTheme.ts'
 import { useAI } from './services/ai.ts'
 import { SCENES, useAppStore, useChatStore } from './stores'
