+++ src/components/AIEditing/storage.ts
@@ -23,9 +23,15 @@ function debounce<T extends unknown[]>(
 // 创建一个防抖的保存函数，500ms 内只执行一次
 const debouncedSaveToServer = debounce(async (_sessionId: string, _content: string) => {
   try {
-    // TODO: 实现 saveDraft API
+    // TODO(priority:medium): 实现草稿自动保存到服务器功能
+    // 需要实现的功能：
+    // 1. 在 AIEditingAPI 中添加 saveDraft(sessionId, content) 方法
+    // 2. 后端实现草稿保存接口 POST /api/drafts/:sessionId
+    // 3. 实现增量保存逻辑（只保存变更部分）
+    // 4. 添加保存失败重试机制
+    // 示例代码：
     // await AIEditingAPI.saveDraft(sessionId, content)
-    // logger.debug('Save draft to server (not implemented):', sessionId, content.substring(0, 50))
+    // logger.debug('Saved draft to server:', sessionId, content.substring(0, 50))
   } catch (error) {
     logger.error('Failed to save content to server:', error)
   }
@@ -51,15 +57,20 @@ export function saveEditorContent(content: string, sessionId?: string): void {
 export async function loadEditorContent(sessionId?: string): Promise<string | null> {
   if (sessionId) {
     try {
-      // TODO: 实现 loadDraft API
-      // 优先从服务器加载
+      // TODO(priority:medium): 实现从服务器加载草稿功能
+      // 需要实现的功能：
+      // 1. 在 AIEditingAPI 中添加 loadDraft(sessionId) 方法
+      // 2. 后端实现草稿加载接口 GET /api/drafts/:sessionId
+      // 3. 添加草稿版本管理（支持历史版本恢复）
+      // 4. 实现离线缓存策略（先从本地加载，再从服务器同步）
+      // 示例代码：
       // const serverContent = await AIEditingAPI.loadDraft(sessionId)
       // if (serverContent) {
+      //   logger.debug('Loaded draft from server:', sessionId)
       //   return serverContent
       // }
-      // logger.debug('Load draft from server (not implemented):', sessionId)
     } catch {
-      // logger.debug('Failed to load content from server')
+      // logger.debug('Failed to load content from server, using local cache')
     }
   }
   return null
