+++ src/components/chat/messages/UserMessage.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import type { Message } from '../../services/database.ts'
+import type { Message } from '@/services/database.ts'
 import { useAppStore } from '@/stores'
+import Markdown from '@/utils/markdown.ts'
 import { CheckIcon, ClipboardDocumentIcon } from '@heroicons/vue/24/outline'
 import { storeToRefs } from 'pinia'
 import { ref } from 'vue'
-import Markdown from '../Markdown.ts'
 
 interface Props {
   message: Message
